requirements.txt
# BMAD Testing Special Agents - Python Dependencies
# This file contains all Python dependencies required to run the BMAD pipeline

# Core HTTP and API clients
requests>=2.31.0
urllib3>=2.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Jira integration (optional - for advanced Jira API features)
# atlassian-python-api>=3.41.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Vector database client (for Qdrant integration)
# qdrant-client>=1.7.0

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Excel processing
openpyxl>=3.1.0
xlrd>=2.0.0

# Document processing
python-docx>=0.8.11

# HTML parsing (for DOM analysis)
beautifulsoup4>=4.12.0
lxml>=4.9.0

# JSON processing and validation
jsonschema>=4.17.0

# Logging and monitoring
colorlog>=6.7.0

# Testing frameworks (for validation)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Type checking and linting
mypy>=1.5.0
black>=23.7.0
flake8>=6.1.0

# Development utilities
tqdm>=4.66.0
click>=8.1.0

# Security
cryptography>=41.0.0

# Utilities
pathlib2>=2.3.0
typing-extensions>=4.7.0

qdrant-client 
pypdf
playwright

openai 
anthropic 
google-generativeai 
python-dotenv 
requests
streamlit>=1.38.0
pillow
